Split lazy into two files
