+++ py-polars/polars/lazy/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
-from .lazy import *
+from .expr import *
+from .frame import *
